+++ spec/recipes/dev_tools_spec.rb
@@ -6,6 +6,11 @@ describe 'bastion::dev_tools' do
   let(:runner) { ChefSpec::SoloRunner.new }
   let(:chef_run) { runner.converge(described_recipe) }
 
+  it 'installs Oracle-flavored Java' do
+    expect(chef_run).to include_recipe('java')
+    expect(chef_run.node['java']['install_flavor']).to eq('oracle')
+  end
+
   it 'installs Git' do
     expect(chef_run).to install_package('git')
   end