+++ test/fixtures/cookbooks/bastion_test/recipes/default.rb
@@ -1,0 +1,8 @@
+# Encoding: UTF-8
+
+service 'rsyslog' do
+  start_command 'service rsyslog start'
+  action :start
+end
+
+include_recipe 'bastion'